Refactor passive voice assessment
