+++ packages/yoastseo/spec/scoring/assessments/readability/PassiveVoiceAssessmentSpec.js
@@ -1,4 +1,4 @@
-import passiveVoiceAssessment from "../../../../src/scoring/assessments/readability/passiveVoiceAssessment.js";
+import PassiveVoiceAssessment from "../../../../src/scoring/assessments/readability/PassiveVoiceAssessment.js";
 import Paper from "../../../../src/values/Paper.js";
 import Factory from "../../../specHelpers/factory.js";
 const i18n = Factory.buildJed();
@@ -9,21 +9,21 @@ import EnglishResearcher from "../../../../src/languageProcessing/languages/en/R
 describe( "An assessment for scoring passive voice.", function() {
 	const paper = new Paper();
 	it( "returns result when the text is empty", function() {
-		const assessment = passiveVoiceAssessment.getResult( paper, Factory.buildMockResearcher( { total: 0, passives: [] } ), i18n );
+		const assessment = new PassiveVoiceAssessment().getResult( paper, Factory.buildMockResearcher( { total: 0, passives: [] } ), i18n );
 		expect( assessment.getScore() ).toBe( 9 );
 		expect( assessment.getText() ).toBe( "<a href='https://yoa.st/34t' target='_blank'>Passive voice</a>: You're using enough" +
 			" active voice. That's great!" );
 	} );
 
 	it( "scores 0 passive sentences - 0%", function() {
-		const assessment = passiveVoiceAssessment.getResult( paper, Factory.buildMockResearcher( { total: 20, passives: [] } ), i18n );
+		const assessment = new PassiveVoiceAssessment().getResult( paper, Factory.buildMockResearcher( { total: 20, passives: [] } ), i18n );
 		expect( assessment.getScore() ).toBe( 9 );
 		expect( assessment.getText() ).toBe( "<a href='https://yoa.st/34t' target='_blank'>Passive voice</a>: You're using enough" +
 			" active voice. That's great!" );
 	} );
 
 	it( "scores 1 passive sentence - 5%", function() {
-		const assessment = passiveVoiceAssessment.getResult( paper, Factory.buildMockResearcher( { total: 20, passives: [ 1 ] } ), i18n );
+		const assessment = new PassiveVoiceAssessment().getResult( paper, Factory.buildMockResearcher( { total: 20, passives: [ 1 ] } ), i18n );
 		expect( assessment.getScore() ).toBe( 9 );
 		expect( assessment.getText() ).toBe( "<a href='https://yoa.st/34t' target='_blank'>Passive voice</a>: You're using enough" +
 			" active voice. That's great!" );
@@ -31,7 +31,7 @@ describe( "An assessment for scoring passive voice.", function() {
 	} );
 
 	it( "scores 2 passive sentences - 10%", function() {
-		const assessment = passiveVoiceAssessment.getResult( paper, Factory.buildMockResearcher( { total: 20, passives: [ 1, 2 ] } ), i18n );
+		const assessment = new PassiveVoiceAssessment().getResult( paper, Factory.buildMockResearcher( { total: 20, passives: [ 1, 2 ] } ), i18n );
 		expect( assessment.getScore() ).toBe( 9 );
 		expect( assessment.getText() ).toBe( "<a href='https://yoa.st/34t' target='_blank'>Passive voice</a>: You're using enough" +
 			" active voice. That's great!" );
@@ -39,7 +39,7 @@ describe( "An assessment for scoring passive voice.", function() {
 	} );
 
 	it( "scores 10 passive sentence - 50%", function() {
-		const assessment = passiveVoiceAssessment.getResult( paper, Factory.buildMockResearcher( { total: 20,
+		const assessment = new PassiveVoiceAssessment().getResult( paper, Factory.buildMockResearcher( { total: 20,
 			passives: [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 ] } ), i18n );
 		expect( assessment.getScore() ).toBe( 3 );
 		expect( assessment.getText() ).toBe( "<a href='https://yoa.st/34t' target='_blank'>Passive voice</a>: 50% of the sentences" +
@@ -49,7 +49,7 @@ describe( "An assessment for scoring passive voice.", function() {
 	} );
 
 	it( "scores 5 passive sentences - 25%", function() {
-		const assessment = passiveVoiceAssessment.getResult( paper, Factory.buildMockResearcher( { total: 20, passives: [ 1, 2, 3, 4, 5 ] } ), i18n );
+		const assessment = new PassiveVoiceAssessment().getResult( paper, Factory.buildMockResearcher( { total: 20, passives: [ 1, 2, 3, 4, 5 ] } ), i18n );
 		expect( assessment.getScore() ).toBe( 3 );
 		expect( assessment.getText() ).toBe( "<a href='https://yoa.st/34t' target='_blank'>Passive voice</a>: 25% of the sentences" +
 			" contain passive voice, which is more than the recommended maximum of 10%. " +
@@ -58,7 +58,7 @@ describe( "An assessment for scoring passive voice.", function() {
 	} );
 
 	it( "scores 5 passive sentences - 13.3%", function() {
-		const assessment = passiveVoiceAssessment.getResult( paper, Factory.buildMockResearcher( { total: 30, passives: [ 1, 2, 3, 4 ] } ), i18n );
+		const assessment = new PassiveVoiceAssessment().getResult( paper, Factory.buildMockResearcher( { total: 30, passives: [ 1, 2, 3, 4 ] } ), i18n );
 		expect( assessment.getScore() ).toBe( 6 );
 		expect( assessment.getText() ).toBe( "<a href='https://yoa.st/34t' target='_blank'>Passive voice</a>: 13.3% of the sentences" +
 			" contain passive voice, which is more than the recommended maximum of 10%. " +
@@ -70,22 +70,22 @@ describe( "An assessment for scoring passive voice.", function() {
 describe( "A test for checking the applicability", function() {
 	it( "returns true for isApplicable for an English paper with text.", function() {
 		const paper = new Paper( "This is a very interesting paper.", { locale: "en_US" } );
-		expect( passiveVoiceAssessment.isApplicable( paper, new EnglishResearcher( paper ) ) ).toBe( true );
+		expect( new PassiveVoiceAssessment().isApplicable( paper, new EnglishResearcher( paper ) ) ).toBe( true );
 	} );
 
 	it( "returns false for isApplicable for an Afrikaans paper with text.", function() {
 		const paper = new Paper( "Hierdie is 'n interessante papier.", { locale: "af_ZA" } );
-		expect( passiveVoiceAssessment.isApplicable( paper, new DefaultResearcher( paper ) ) ).toBe( false );
+		expect( new PassiveVoiceAssessment().isApplicable( paper, new DefaultResearcher( paper ) ) ).toBe( false );
 	} );
 
 	it( "returns false for isApplicable for an English paper without text.", function() {
 		const paper = new Paper( "", { locale: "en_US" } );
-		expect( passiveVoiceAssessment.isApplicable( paper, new EnglishResearcher( paper ) ) ).toBe( false );
+		expect( new PassiveVoiceAssessment().isApplicable( paper, new EnglishResearcher( paper ) ) ).toBe( false );
 	} );
 
 	it( "returns false for isApplicable for an Afrikaans paper without text.", function() {
 		const paper = new Paper( "", { locale: "af_ZA" } );
-		expect( passiveVoiceAssessment.isApplicable( paper, new DefaultResearcher( paper ) ) ).toBe( false );
+		expect( new PassiveVoiceAssessment().isApplicable( paper, new DefaultResearcher( paper ) ) ).toBe( false );
 	} );
 } );
 
@@ -97,13 +97,13 @@ describe( "A test for marking passive sentences", function() {
 			new Mark( { original: "A very interesting paper has been written.",
 				marked: "<yoastmark class='yoast-text-mark'>A very interesting paper has been written.</yoastmark>" } ),
 		];
-		expect( passiveVoiceAssessment.getMarks( paper, passiveVoice ) ).toEqual( expected );
+		expect( new PassiveVoiceAssessment().getMarks( paper, passiveVoice ) ).toEqual( expected );
 	} );
 
 	it( "returns no markers for active sentences", function() {
 		const paper = new Paper( "This is a very interesting paper." );
 		const passiveVoice = Factory.buildMockResearcher( [ { total: 0, passives: [] } ] );
 		const expected = [];
-		expect( passiveVoiceAssessment.getMarks( paper, passiveVoice ) ).toEqual( expected );
+		expect( new PassiveVoiceAssessment().getMarks( paper, passiveVoice ) ).toEqual( expected );
 	} );
 } );
