+++ packages/yoastseo/spec/scoring/assessments/readability/TransitionWordsAssessmentSpec.js
