+++ packages/yoastseo/src/scoring/assessments/index.js
@@ -1,7 +1,7 @@
 import FleschReadingEaseAssessment from "./readability/fleschReadingEaseAssessment";
 import ListAssessment from "./readability/ListAssessment";
 import ParagraphTooLongAssessment from "./readability/ParagraphTooLongAssessment";
-import PassiveVoiceAssessment from "./readability/passiveVoiceAssessment";
+import PassiveVoiceAssessment from "./readability/PassiveVoiceAssessment";
 import SentenceBeginningsAssessment from "./readability/sentenceBeginningsAssessment";
 import SentenceLengthInTextAssessment from "./readability/sentenceLengthInTextAssessment";
 import SubheadingDistributionTooLongAssessment from "./readability/subheadingDistributionTooLongAssessment";
