+++ packages/yoastseo/src/scoring/contentAssessor.js
@@ -4,7 +4,7 @@ import ParagraphTooLong from "./assessments/readability/ParagraphTooLongAssessme
 import SentenceLengthInText from "./assessments/readability/sentenceLengthInTextAssessment.js";
 import SubheadingDistributionTooLong from "./assessments/readability/subheadingDistributionTooLongAssessment.js";
 import TransitionWords from "./assessments/readability/TransitionWordsAssessment.js";
-import passiveVoice from "./assessments/readability/passiveVoiceAssessment.js";
+import passiveVoice from "./assessments/readability/PassiveVoiceAssessment.js";
 import sentenceBeginnings from "./assessments/readability/sentenceBeginningsAssessment.js";
 import textPresence from "./assessments/readability/textPresenceAssessment.js";
