+++ packages/yoastseo/src/scoring/productPages/contentAssessor.js
@@ -4,7 +4,7 @@ import ParagraphTooLong from "../assessments/readability/ParagraphTooLongAssessm
 import SentenceLengthInText from "../assessments/readability/sentenceLengthInTextAssessment.js";
 import SubheadingDistributionTooLong from "../assessments/readability/subheadingDistributionTooLongAssessment.js";
 import TransitionWords from "../assessments/readability/TransitionWordsAssessment.js";
-import passiveVoice from "../assessments/readability/passiveVoiceAssessment.js";
+import passiveVoice from "../assessments/readability/PassiveVoiceAssessment.js";
 import textPresence from "../assessments/readability/textPresenceAssessment.js";
 import ListsPresence from "../assessments/readability/ListAssessment.js";
